+++ src/solver.cpp
@@ -6,6 +6,6 @@ Solver::~Solver() { delete guesser; }
 
 int Solver::score() { return this->guesser->getGuessCount(); }
 
-ReturnType Solver::guess(unsigned int x, unsigned int y, Board &board) {
+ReturnType Solver::guess(size_t x, size_t y, Board &board) {
     return this->guesser->guess(x, y, board);
 }
